% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_create_iam_role.R
\name{aws_create_iam_role}
\alias{aws_create_iam_role}
\title{aws_create_iam_role}
\usage{
aws_create_iam_role(
  url = NULL,
  token = NULL,
  auth_type = "iam",
  role_name = NULL,
  bound_iam_principal_arn = NULL,
  policies = NULL
)
}
\arguments{
\item{url}{URL of the HashiCorp Vault instance.}

\item{token}{token from user/github/approle/etc.... registered in Vault.}

\item{auth_type}{authentication type for the role you are creating e.g. iam.}

\item{role_name}{name of the role you are creating.}

\item{bound_iam_principal_arn}{arn of the iam role that you created in AWS.}

\item{policies}{policy or a list of policies in Vault to associate the IAM role with.}
}
\value{
Return's a list of aws roles present in your Vault instance.
}
\description{
Vault: Creates the Vault AWS IAM role to interact with the IAM role in your AWS account.
}
\details{
This function creates the Vault AWS IAM Role and associates it with the IAM Role that has already been created in AWS.
}
\examples{
\dontrun{
aws_create_iam_role(url=NULL,token=NULL,auth_type="iam",role_name=NULL,bound_iam_principal_arn=NULL,policies=NULL)

}

}
\references{
\href{https://developer.hashicorp.com/vault/api-docs/v1.16.x/auth/aws#create-update-role}{Vault AWS Create Role Documentation} \if{html}{\out{<br/>}}
}
\keyword{aws_create_iam_role}
