% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_list_roles.R
\name{aws_list_roles}
\alias{aws_list_roles}
\title{aws_list_roles}
\usage{
aws_list_roles(url = NULL, token = NULL)
}
\arguments{
\item{url}{URL of the HashiCorp Vault instance.}

\item{token}{token from user/github/approle/etc.... registered in Vault.}
}
\value{
Return's a list of aws roles present in your Vault instance.
}
\description{
Vault: Lists/Gets the AWS role present in your Vault instance.
}
\details{
This function lists the AWS roles present in your Vault instance.
}
\examples{

```r
aws_list_roles(url="vault-url.com",token=token)

```

}
\references{
\href{https://developer.hashicorp.com/vault/api-docs/v1.16.x/auth/aws#list-roles}{Vault AWS LIST Roles Documentation} \if{html}{\out{<br/>}}
\href{https://httr.r-lib.org/reference/VERB.html}{HTTR Package VERB Function} \if{html}{\out{<br/>}}
\if{html}{\out{<a href="https://www.r-project.org/" target="_blank">}}Open R Project in a new tab\if{html}{\out{</a>}}
}
\keyword{aws_list_roles}
