% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_rotate_root_creds.R
\name{aws_rotate_root_creds}
\alias{aws_rotate_root_creds}
\title{aws_rotate_root_creds}
\usage{
aws_rotate_root_creds(url = NULL, token = NULL)
}
\arguments{
\item{url}{URL of the HashiCorp Vault instance.}

\item{token}{token (preferrably) the root token to rotate the AWS credentials being used by Vault.}
}
\value{
Return's the access key that has been rotated in AWS.
}
\description{
Vault: Rotates the root credentials used in AWS by Vault.
}
\details{
This function rotates the root IAM User credentials used in AWS by Vault.
}
\examples{

\dontrun{  aws_rotate_root_creds(url,token)

}

}
\keyword{aws_rotate_root_creds}
