% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkout.R
\name{checkout}
\alias{checkout}
\title{Function to checkout file from vault}
\usage{
checkout(vault, item, openFile = TRUE)
}
\arguments{
\item{vault}{a vault object}

\item{item}{the item to checkout as a character string}

\item{openFile}{a toggle specifying whether to open the file after download in R studio}
}
\description{
Function to checkout file from vault
}
