% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{checkVaultAccess}
\alias{checkVaultAccess}
\title{Check if have access to vault}
\usage{
checkVaultAccess(repo)
}
\arguments{
\item{repo}{a string set up as \if{html}{\out{<org>}}/\if{html}{\out{<repo>}}}
}
\description{
Check if have access to vault
}
