% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_aws_root_creds.R
\name{config_aws_root_creds}
\alias{config_aws_root_creds}
\title{config_aws_root_creds}
\usage{
config_aws_root_creds(
  url = NULL,
  token = NULL,
  aws_key = NULL,
  aws_secret = NULL,
  aws_region = NULL
)
}
\arguments{
\item{url}{URL of the Hashicorp Vault instance.}

\item{token}{token from user/github/approle/etc.... registered in Vault.}

\item{aws_key}{AWS Access Key ID from IAM User registered in AWS.}

\item{aws_secret}{AWS Secret Access Key from IAM User registered in AWS.}

\item{aws_region}{AWS Region where IAM User is registered in}
}
\value{
writes the AWS Accesss Key ID, AWS Secret Access Key, Region used by the registered IAM User in AWS to Vault.
}
\description{
Vault: Registers the AWS User credentials in Vault. Note: Please do not register the ROOT User tokens from your AWS instance.
}
\details{
This function registers the AWS User credentials in Vault used to get the Access Key ID and Secret Access Key for Users and roles in AWS.
}
\note{
Examples of role policy and trust policy required in AWS:
'\preformatted{Example Trust Policy required in AWS: 
'{
'  "Version": "2012-10-17",
'  "Statement": [
'    {
'      "Effect": "Allow",
'      "Principal": {
'        "AWS": "arn:aws:iam::ACCOUNT-ID-WITHOUT-HYPHENS:user/VAULT-AWS-ROOT-CONFIG-USER-NAME"
'      },
'      "Action": "sts:AssumeRole"
'    }
'  ]
'}
'}

'\preformatted{Example IAM User Policy required in AWS: 
'{
'"Version": "2012-10-17",
'"Statement": {
'  "Effect": "Allow",
'  "Action": "sts:AssumeRole",
'  "Resource": "arn:aws:iam::ACCOUNT-ID-WITHOUT-HYPHENS:role/RoleNameToAssume"
'}
'}
'}
}
\examples{

\dontrun{  config_aws_root_creds(url,token,aws_key,aws_secret,aws_region)

}

}
\references{
\url{https://www.vaultproject.io/docs/secrets/aws}
}
\keyword{config_aws_root_creds}
