% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_approle_token.R
\name{get_approle_token}
\alias{get_approle_token}
\title{get_approle_token}
\usage{
get_approle_token(url, role_id, secret_id)
}
\arguments{
\item{url}{url of the Hashicorp Vault instance.}

\item{role_id}{role_id of the approle in Vault.}

\item{secret_id}{secret_id of the approle in Vault.}
}
\value{
Return's the user token that allows an approle to query secrets in Vault.
}
\description{
Vault: Gets the Vault Data
}
\details{
This function gets secrets or data from the vault instance and returns them inside a dataframe.
You will need the Hashicorp Vault url you are using including the full path to the secret e.g. (https://url.vault.com/v1/secret/things)
and lastly you will need a token that allows you to interact with Vault via the API. You can pass a list of secrets at once use 'list(one="two",three="four) and so on and so forth. This will allow you to send several secrets to Vault at once.
}
\examples{

\dontrun{  get_approle_token(url,role_id,secret_id)

}

}
\keyword{get_approle_token}
