% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aws_creds.R
\name{get_aws_creds}
\alias{get_aws_creds}
\title{get_aws_creds}
\usage{
get_aws_creds(url = NULL, token = NULL, role_name = NULL)
}
\arguments{
\item{url}{URL of the Hashicorp Vault instance.}

\item{token}{token from user/github/approle/etc.... registered in Vault.}

\item{role_name}{role_name for role registered in Vault}
}
\value{
Return's the user token that allows an individual to query secrets in Vault.
}
\description{
Vault: Gets AWS creds from Vault.
}
\details{
This function gets the AWS creds from the role/IAM User registered in Vault.
}
\examples{

\dontrun{  get_aws_creds(url,token,role_name)

}

}
\keyword{get_aws_creds}
