% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aws_role_info.R
\name{get_aws_role_info}
\alias{get_aws_role_info}
\title{get_aws_role_info}
\usage{
get_aws_role_info(url = NULL, token = NULL, role_name = NULL)
}
\arguments{
\item{url}{URL of the Hashicorp Vault instance.}

\item{token}{token from user/github/approle/etc.... registered in Vault.}

\item{role_name}{role_name for role registered in Vault}
}
\value{
Return's the user token that allows an individual to query secrets in Vault.
}
\description{
Vault: Gets the AWS role info from Vault.
}
\details{
This function gets the AWS role info from Vault for the role registered in your Vault instance.
}
\examples{

\dontrun{  get_aws_role_info(url,token,role_name)

}

}
\keyword{get_aws_role_info}
