% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aws_sts.R
\name{get_aws_sts}
\alias{get_aws_sts}
\title{get_aws_sts}
\usage{
get_aws_sts(url = NULL, token = NULL, role_name = NULL)
}
\arguments{
\item{url}{url of the Hashicorp Vault instance.}

\item{token}{token from user/github/approle/etc.... registered in Vault.}

\item{role_name}{role_name for role registered in Vault}
}
\value{
Return's the user token that allows an individual to query secrets in Vault.
}
\description{
Vault: Gets the AWS STS credentials from Vault.
}
\details{
This function gets AWS STS credentials from the registered role in your Vault instance.
}
\examples{

\dontrun{  get_aws_sts(url,token,role_name)

}

}
\keyword{get_aws_creds}
