% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_role_id.R
\name{get_role_id}
\alias{get_role_id}
\title{get_role_id}
\usage{
get_role_id(url, role_name, token)
}
\arguments{
\item{url}{URL of the Hashicorp Vault instance.}

\item{role_name}{name of the role in your Vault instance.}

\item{token}{token from your loging role such as user/pass method or github/auth method.}
}
\value{
Return's the user/github/etc... token that allows an individual to query secrets in Vault via an approle.
}
\description{
Vault: Gets the Approle ID from Vault.
}
\details{
This function obtains the approle ID from Vault. The user can then utilize the approle ID retrieved to query secrets from Vault via the approle method. 
You will need the URL of the Hashicorp Vault you are using and your user/github/etc... token to retrieve the approle id.
}
\examples{

\dontrun{  get_role_id(url='vautl.url',role_name='jenkins',token='12345abcdef')

}

}
\keyword{get_role_id}
