% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_secret_id.R
\name{get_secret_id}
\alias{get_secret_id}
\title{get_secret_id}
\usage{
get_secret_id(url, role_name, token)
}
\arguments{
\item{url}{url of the Hashicorp Vault instance.}

\item{role_name}{name of the role in your Vault instance.}

\item{token}{token from your user/github/etc... login method}
}
\value{
Return's the secret id that allows an individual to query secrets in Vault via the approle method.
}
\description{
Vault: Gets the Vault Data
}
\details{
This function obtains the token for the github user from Vault. The user can then utilize the token retrieved to query secrets from Vault. 
You will need the url of the Hashicorp Vault you are using and your PAT from github. You should have already created an organization in github
to utilize this login path in Vault.
}
\keyword{get_secret_id}
