% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{get_token}
\alias{get_token}
\alias{data}
\title{get_token}
\usage{
get_token(url, user, pass)
}
\arguments{
\item{url}{url of the Hashicorp Vault instance.}

\item{user}{username registered in Vault.}

\item{pass}{password for username registered in Vault}
}
\value{
Return's the user token that allows an individual to query results in Vault.
}
\description{
Vault: Gets the Vault Data
}
\details{
This function gets secrets or data from the vault instance and returns them inside a dataframe.
You will need the Hashicorp Vault url you are using including the full path to the secret e.g. (https://url.vault.com/v1/secret/things)
and lastly you will need a token that allows you to interact with Vault via the API. You can pass a list of secrets at once use 'list(one="two",three="four) and so on and so forth. This will allow you to send several secrets to Vault at once.
}
\examples{

\dontrun{  get_token(url,user,pass)

}

}
\keyword{get_token}
