% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_token.R
\name{get_user_token}
\alias{get_user_token}
\title{get_user_token}
\usage{
get_user_token(url, user, pass)
}
\arguments{
\item{url}{URL of the Hashicorp Vault instance.}

\item{user}{username registered in Vault.}

\item{pass}{password for username registered in Vault}
}
\value{
Return's the user token that allows an individual to query secrets in Vault.
}
\description{
Vault: Gets the Vault Data
}
\details{
This function gets the user token from Vault for the user name registered in your Vault instance registered at the auth/userpass path.
}
\examples{

\dontrun{  get_user_token(url,user,pass)

}

}
\keyword{get_token}
