% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_get_token.R
\name{github_get_token}
\alias{github_get_token}
\alias{data}
\title{github_get_token}
\usage{
github_get_token(url, github_token)
}
\arguments{
\item{url}{url of the Hashicorp Vault instance.}

\item{github_token}{personal authentication token or PAT from your github repo.}
}
\value{
Return's the user token that allows an individual to query secrets in Vault.
}
\description{
Vault: Gets the Vault Data
}
\details{
This function obtains the token for the github user from Vault. The user can then utilize the token retrieved to query secrets from Vault. 
You will need the url of the Hashicorp Vault you are using and your PAT from github. You should have already created an organization in github
to utilize this login path in Vault.
}
\keyword{github_get_token}
