% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_aws_roles.R
\name{list_aws_roles}
\alias{list_aws_roles}
\title{list_aws_roles}
\usage{
list_aws_roles(url = NULL, token = NULL)
}
\arguments{
\item{url}{URL of the Hashicorp Vault instance.}

\item{token}{token from user/github/approle/etc.... registered in Vault.}
}
\value{
Return's the AWS roles that are registered in Vault.
}
\description{
Vault: Lists the AWS roles registered in Vault.
}
\details{
This function lists the AWS roles that are registered in Vault.
}
\examples{

\dontrun{  list_aws_roles(url,token)

}

}
\keyword{list_aws_roles}
