% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{listDetails}
\alias{listDetails}
\title{List details of vault}
\usage{
listDetails(vault)
}
\arguments{
\item{vault}{a vault object}
}
\value{
a detailed tibble with information about the vault. This takes longer to load
}
\description{
List details of vault
}
