% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkout.R
\name{mapCheckout}
\alias{mapCheckout}
\title{Function to checkout multiple files from vault}
\usage{
mapCheckout(vault, items)
}
\arguments{
\item{vault}{a vault object}

\item{items}{the items to checkout as a character vector}
}
\description{
Function to checkout multiple files from vault
}
