% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{news}
\alias{news}
\title{Get software news for the vault item in Rstudio viewer}
\usage{
news(vault, item)
}
\arguments{
\item{vault}{a vault object}

\item{item}{the item in the vault to preview}
}
\value{
opens r studio viewer to preview vault item news
}
\description{
Get software news for the vault item in Rstudio viewer
}
