% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{preview}
\alias{preview}
\title{Preview file in Rstudio viewer}
\usage{
preview(vault, item)
}
\arguments{
\item{vault}{a vault object}

\item{item}{the item in the vault to preview}
}
\description{
Preview file in Rstudio viewer
}
