% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{vault}
\alias{vault}
\title{Create a vault object to use}
\usage{
vault(repo_string, savePath = here::here())
}
\arguments{
\item{repo_string}{a string specified as \if{html}{\out{<org>}}/\if{html}{\out{<repo>}}}

\item{savePath}{a path to use for saving items from vault, defaults to project directory}
}
\description{
Create a vault object to use
}
