% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{vaultContents}
\alias{vaultContents}
\title{Look up contents of vault}
\usage{
vaultContents(owner, repo)
}
\arguments{
\item{owner}{name of organization}

\item{repo}{name of repository}
}
\description{
Look up contents of vault
}
