% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_approle_token.R
\name{get_approle_token}
\alias{get_approle_token}
\title{get_approle_token}
\usage{
get_approle_token(url, role_id, secret_id)
}
\arguments{
\item{url}{URL of the Hashicorp Vault instance.}

\item{role_id}{role_id of the approle in Vault.}

\item{secret_id}{secret_id of the approle in Vault.}
}
\value{
Return's the user token that allows an approle to query secrets in Vault.
}
\description{
Vault: Gets the approle token from Vault.
}
\details{
This function gets the approle token from the vault instance. You can then utilize the approle token to query data from vault.
}
\examples{

\dontrun{  get_approle_token(url,role_id,secret_id)

}

}
\keyword{get_approle_token}
