% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aws_creds.R
\name{get_aws_creds}
\alias{get_aws_creds}
\title{get_aws_creds}
\usage{
get_aws_creds(url = NULL, role = NULL, token = NULL)
}
\arguments{
\item{url}{url of the Hashicorp Vault instance.}

\item{role}{role to get secret key and secret access}

\item{token}{token to query Vault}
}
\value{
Return's the user token that allows an individual to query secrets in Vault.
}
\description{
Vault: Gets the Vault Data
}
\details{
This function gets AWS user creds via Vault to be used to connect to AWS services such as s3 buckets, ec2 instances, etc...
}
\examples{

\dontrun{  get_aws_creds(url,role,token)

}

}
\keyword{get_token}
