% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_role_id.R
\name{get_role_id}
\alias{get_role_id}
\title{get_role_id}
\usage{
get_role_id(url, role_name, token)
}
\arguments{
\item{url}{url of the Hashicorp Vault instance.}

\item{role_name}{name of the role in your Vault instance.}

\item{token}{token from your loging role such as user/pass method or github/auth method.}
}
\value{
Return's the user/github/etc... token that allows an individual to query secrets in Vault via an approle.
}
\description{
Vault: Gets the Vault Data
}
\details{
This function obtains the token for the user/github/etc... user from Vault. The user can then utilize the token retrieved to query secrets from Vault via the approle method. 
You will need the url of the Hashicorp Vault you are using and your user/github/etc... token to retrieve the approle id.
}
\examples{

\dontrun{  get_role_id(url='vautl.url',role_name='jenkins',github_token='12345abcdef')

}

}
\keyword{get_role_id}
