% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_secret_id.R
\name{get_secret_id}
\alias{get_secret_id}
\title{get_secret_id}
\usage{
get_secret_id(url, role_name, token)
}
\arguments{
\item{url}{URL of the Hashicorp Vault instance.}

\item{role_name}{name of the role in your Vault instance.}

\item{token}{token from your user/github/etc... login method}
}
\value{
Return's the secret id that allows an individual to query secrets in Vault via the approle method.
}
\description{
Vault: Gets the secret id for the approle registered in vault e.g.(jenkins).
}
\details{
This function obtains the secret id for the approle from Vault. The user can then utilize the secret id retrieved to obtain the approle token from Vault. 
You will need the URL of the Hashicorp Vault you are using the role name in vault and a token from an auth path in Vault.
}
\examples{
\dontrun{  get_secret_id(url='vautl.url',role_name='jenkins',token='12345abcdef')

}

}
\keyword{get_secret_id}
