% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{vaultContents}
\alias{vaultContents}
\title{Look up contents of vault}
\usage{
vaultContents(repo)
}
\arguments{
\item{repo}{a string set up as \if{html}{\out{<org>}}/\if{html}{\out{<repo>}}}
}
\description{
Look up contents of vault
}
